// packages/runtime-vapor/js-reactive/types.d.ts
export type ObjectLike = Record<PropertyKey, unknown>
export interface HookHost extends ObjectLike {
  __hooks?: unknown
}
export interface HookContainer extends ObjectLike {
  states: unknown[]
  index: number
  __forcedIndex?: number
  __idMap?: Map<unknown, unknown>
}
export interface HookFrame {
  instance: unknown
  hooks?: HookContainer
}
export interface HookCarrier {
  getCurrentInstance(): unknown
  renderHooks<T>(instance: unknown, render: () => T): T
}
export interface HookContext extends HookCarrier {
  setCurrentInstance(instance: unknown): void
  vaporWithHookId<T>(id: unknown, runner: () => T): T
  withHookSlot<T>(factory: () => T): T
}
// ...
export type EqualityComparator<T> = (previous: T, next: T) => boolean
export interface EqualityOptions<T> {
  equals?: EqualityComparator<T>
}
export interface ReactiveOptions<T> extends EqualityOptions<T> {
  readonly?: boolean
  shallow?: boolean
}
export interface RefLike<T> extends ObjectLike {
  value: T
  __rue_ref__?: true
}
// ...
export type StateKind = 'reactive' | 'ref' | 'signal'
export interface StateOptions<T> extends EqualityOptions<T> {
  kind?: StateKind
}
export interface ResolvedStateOptions {
  equals?: EqualityComparator<unknown>
  kind: StateKind
}
export interface SignalHandle<T> extends KernelSignalHandle<T>, ObjectLike {}
export interface RefState<T> extends RefLike<T> {}
export type ReactiveState<T> = T extends object ? T : RefState<T>
export type StateInitializer<T> = T | (() => T)
export type StateUpdate<TState, TValue> = TValue | ((state: TState) => TValue | void)
export type StateSetter<TState, TValue> = (update: StateUpdate<TState, TValue>) => void
export type StateTuple<TState, TValue> = [TState, StateSetter<TState, TValue>]
export interface StateSlot {
  created: boolean
  state: unknown
  __wrapped__?: boolean
}
export type HookDependencies = readonly unknown[]
export interface MemoSlot<T> {
  value: T | undefined
  deps: unknown
}
export interface RefSlot<T> {
  current: T
}
export interface StateValueHooks {
  createReactive(initial: unknown, options?: EqualityOptions<unknown>): unknown
  isReactive(value: unknown): boolean
}
export interface CreateStateHooksOptions {
  context: HookContext
  reactiveRuntime: unknown
  values: StateValueHooks
}